renderer-ssr: add missing plugins, split tests into multiple files
