+++ packages/public/renderer-ssr/__tests-ssr__/plugin-important-statement.ts
@@ -1,0 +1,23 @@
+import { plugins } from '@edtr-io/internal__fixtures'
+import { StateTypeSerializedType } from '@edtr-io/internal__plugin-state'
+import { ImportantStatementState } from '@edtr-io/plugin-important-statement'
+
+import { render } from '../src'
+
+test('Important statement plugin', () => {
+  const state: {
+    plugin: string
+    state: StateTypeSerializedType<ImportantStatementState>
+  } = {
+    plugin: 'importantStatement',
+    state: {
+      plugin: 'anchor',
+      state: 'foo'
+    }
+  }
+  const { html } = render({
+    state,
+    plugins
+  })
+  expect(html).toContain('foo')
+})
